requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.8.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-filter==2.4.0
django-htmx==1.6.0
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
django-restframework==0.0.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
Faker==11.3.0
idna==3.3
lxml==4.8.0
Markdown==3.3.4
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.50
Pillow==8.4.0
Pint==0.18
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.9
Werkzeug==2.0.2
